+++ config/routes.rb
@@ -6,6 +6,16 @@ require 'sidekiq-scheduler/web'
 Rails.application.routes.draw do
   namespace :backoffice do
     # TODO: spec this
+    # https://github.com/mperham/sidekiq/wiki/Monitoring#rails-http-basic-auth-from-routes
+    Sidekiq::Web.use Rack::Auth::Basic do |username, password|
+      # Protect against timing attacks:
+      # - See https://codahale.com/a-lesson-in-timing-attacks/
+      # - See https://thisdata.com/blog/timing-attacks-against-string-comparison/
+      # - Use & (do not use &&) so that it doesn't short circuit.
+      # - Use digests to stop length information leaking (see also ActiveSupport::SecurityUtils.variable_size_secure_compare)
+      ActiveSupport::SecurityUtils.secure_compare(::Digest::SHA256.hexdigest(username), ::Digest::SHA256.hexdigest(ENV["SIDEKIQ_USERNAME"])) &
+        ActiveSupport::SecurityUtils.secure_compare(::Digest::SHA256.hexdigest(password), ::Digest::SHA256.hexdigest(ENV["SIDEKIQ_PASSWORD"]))
+      end
     mount Sidekiq::Web, at: 'sidekiq'
   end